Navigator and Content provider
The navigator loads the page content basing on the url.
Furthermore, override the internal links behaviour loading the
chosen content.

The content provider is a middleware between the navigator and the
WordPress utilities: it queries the wp functions end return a DOM
node to be setted as content by the navigator.